fix(coding-agent): clear goal state on headless session switch + dedupe duration formatter
Address PR #8225 review (roboomp):

- restore() now clears the in-memory #goalModeState and resets the
  continuation flags on every early-return branch (disabled / non-goal /
  invalid-goal) via a new #clearGoalModeState() helper. Previously a
  headless (ACP/RPC) goal→non-goal session switch left the prior goal live
  because the headless reconciler drives switches through restore() alone
  (the TUI clears via #clearTransientModeState first), injecting goal
  prompts and firing continuations in a goal-less session. Adds a
  controller-level regression test.
- Extract the coarse formatDuration into a theme-free
  slash-commands/helpers/duration.ts (re-exported from helpers/format.ts,
  which keeps renderAsciiBar + its theme imports) and drop the forked
  formatGoalDuration from the controller; output format unchanged.
