+++ packages/coding-agent/src/slash-commands/helpers/format.ts
@@ -1,17 +1,7 @@
 import { shimmerText } from "../../modes/theme/shimmer";
 import { theme as currentTheme, type Theme } from "../../modes/theme/theme";
 
-/** Format a millisecond duration as a coarse-grained human label. */
-export function formatDuration(ms: number): string {
-	const seconds = Math.max(0, Math.round(ms / 1000));
-	if (seconds < 60) return `${seconds}s`;
-	const minutes = Math.round(seconds / 60);
-	if (minutes < 60) return `${minutes}m`;
-	const hours = Math.round(minutes / 60);
-	if (hours < 48) return `${hours}h`;
-	const days = Math.round(hours / 24);
-	return `${days}d`;
-}
+export { formatDuration } from "./duration";
 
 type ProgressBarTheme = Pick<Theme, "bold" | "fg" | "getFgAnsi">;
